# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        '200':
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessage'
      responses:
        '200':
          description: The created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    patch:
      summary: 'Update message by id'
      tags: 
        - messages
      requestBody:
          description: Message updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: The updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      summary: 'Delete message by id'
      tags: 
        - messages
      requestBody:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageID"
      responses:
        '200':
          description: delete message
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: delete message
          content:
            text/plain:
              schema:
                type: string
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        400:
          description: Error
          content:
            text/plain:
              schema:
                type: string
        '200':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      summary: 'Update user by id'
      tags: 
        - users
      requestBody:
          description: user updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUser'
      responses:
        400:
          description: error
          content:
            text/plain:
              schema:
                type: string
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: 'Delete user by id'
      tags: 
        - users
      requestBody:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserID"
      responses:
        '200':
          description: delete user
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: delete user
          content:
            text/plain:
              schema:
                type: string              
            
                

              
components:
  schemas:

# Messages

    NewMessage:
      type: object
      properties:
        message:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint
        message:
          type: string
    MessageID:
      type: object
      properties:
        id:
          type: integer
          format: uint

# Users

    NewUser:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    UpdateUser:
      oneOf:
        - $ref: '#/components/schemas/UserEmail'
        - $ref: '#/components/schemas/UserPassword'
        - $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string
    UserEmail:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
    UserPassword:
      type: object
      properties:
        id:
          type: integer
          format: uint
        password:
          type: string
    UserID:
      type: object
      properties:
        id:
          type: integer
          format: uint